streamlit_app_v.0.2/requirements.txt
streamlit
plotly
Pillow
pandas
numpy
scipy